Se cambiaron las entradas de tipo entero a tipo flotante para mayor precisión en: Ejercicio1, Ejercicio2, Ejercicio3, Ejercicio4, Ejercicio5, Ejercicio6, Ejercicio7, Ejercicio8, Ejercicio9, Ejercicio10, Ejercicio11, Ejercicio12, Ejercicio13, Ejercicio14, Ejercicio15, Ejercicio16, Ejercicio17, Ejercicio18, Ejercicio19, Ejercicio20, Ejercicio21, Ejercicio22, Ejercicio23, Ejercicio24, Ejercicio25, Ejercicio26, Ejercicio27, Ejercicio29 y Ejercicio30.
Se actualizaron los print para usar f-strings y mejorar la legibilidad.

Se añadió manejo de errores para entradas inválidas en Ejercicio20 y Ejercicio27.

Se ajustaron los cálculos de área y perímetro en Ejercicio2 y Ejercicio3.

Se mejoraron los mensajes de solicitud al usuario en varios ejercicios para mayor claridad.

+++ Ejercicios Diagrama/Ejercicio18.py
@@ -1,5 +1,5 @@
-n1 = int(input("Escribe el primer número: "))
-n2 = int(input("Escribe el segundo número: "))
+n1 = float(input("Escribe el primer número: "))
+n2 = float(input("Escribe el segundo número: "))
 if n1 != n2:
     print ("El número mayor es", max(n1, n2))
 else:
