+++ Ejercicios Diagrama/Ejercicio23.py
@@ -4,7 +4,7 @@ neg, pos = 0, 0
 num = -1
 
 while num != 0:
-    num = int(input("Escribe un número entero: "))
+    num = float(input("Escribe un número entero: "))
         
     if num < 0:
         neg +=1
@@ -21,7 +21,7 @@ print("Números positivos:", pos, "\n"
 neg, pos = 0, 0
 
 while True:
-    num = int(input("Escribe un número entero: "))
+    num = float(input("Escribe un número entero: "))
         
     if num < 0:
         neg +=1
