+++ Ejercicios Parte 1/Ejercicio11.py
@@ -1,3 +1,3 @@
-n1 = int(input("Escribe el primer número: "))
-n2 = int(input("Escribe el segundo número: "))
+n1 = float(input("Escribe el primer número: "))
+n2 = float(input("Escribe el segundo número: "))
 print("El mayor es", n1 if n1 >= n2 else n2)
