+++ Ejercicios Parte 1/Ejercicio12.py
@@ -1,2 +1,2 @@
-n1 = int(input("Escribe un número: "))
+n1 = float(input("Escribe un número: "))
 print("El número es positivo." if n1 >= 0 else "El número es negativo.")
