+++ Ejercicios Parte 1/Ejercicio13.py
@@ -1,5 +1,5 @@
-n1 = int(input("Escribe el primer número: "))
-n2 = int(input("Escribe el segundo número: "))
+n1 = float(input("Escribe el primer número: "))
+n2 = float(input("Escribe el segundo número: "))
 
 if n1 < n2:
     print(n1, "<", n2)
